Feature: Moved the binary and platform resolving logic to the DownloadUrlResolver
